chore: rename compound_keys to composite_keys
